refactor(water): two products that share a noun and no method
water.py becomes terra/water/indices.py and flood.py becomes
terra/flood/envelope.py: two slices, not one package named water holding both.

They have nothing in common but the word. terra/water reads reflectance -- a
spectral index thresholded per date, with no trained model and no fixed legend,
so none of the classifier's domain limitation applies to it. terra/flood reads
terrain -- HAND over a merged DEM window, flow accumulation, and the
disagreement between four elevation products -- and touches no imagery at all.
A package holding both would be named for the noun their outputs share, which
is the mistake a package named util makes with a different word.

The rest of the application already models them as two. Two actions, `water`
and `flood_envelope`; two run kinds in internal/store/store.go; types_flood.go
of its own; separate panels. The sidecar now says the same thing.

The layers contract gains both beside terra.energy, and a third contract is now
worth having: the three products must not import each other. It passes today
and it is the one that stops a slice from quietly growing a dependency on
another slice's internals.

Verification: pytest 515 passed, 8 skipped. ruff check clean, three contracts
kept, gofmt, go vet and go test -race clean. A ping through the entry point
answers.
