+++ sidecar/pyproject.toml
@@ -72,11 +72,16 @@ layers = [
     "terra.cli",
     "terra.registry",
     "terra.actions",
-    "terra.energy",
+    "terra.energy | terra.water | terra.flood",
     "terra.imagery | terra.terrain | terra.sun",
     "terra.stac",
     "terra.protocol",
 ]
+
+[[tool.importlinter.contracts]]
+name = "Products do not reach into each other"
+type = "independence"
+modules = ["terra.energy", "terra.water", "terra.flood"]
 
 [[tool.importlinter.contracts]]
 name = "The sun service knows nothing about the products that read it"
