+++ sidecar/terra/flood/__init__.py
@@ -1,0 +1,9 @@
+"""
+The flood envelope: how much of a HAND extent is terrain, and how much is the
+choice of DEM.
+
+A product about the disagreement between elevation products, not about water
+observed from orbit. It reads terra.terrain and reaches no imagery at all,
+which is why it is a slice of its own rather than a module beside the spectral
+water indices.
+"""
