+++ sidecar/terra/water/__init__.py
@@ -1,0 +1,8 @@
+"""
+Surface water from what the satellite measured, and nothing else.
+
+A thresholded spectral index over a Sentinel-2 series: no trained model, no
+fixed legend, and therefore none of the domain limitation the classifier
+carries. It is a different product from the flood envelope, which shares the
+word water and no method: that one reads terrain and this one reads reflectance.
+"""
